Log human-readable traces in tests
