+++ internal/printer/colored/colored.go
